(PC-18553) feat(ui): divide animationHelpers and move to ui/animations/helpers
