+++ src/features/search/components/SearchResults/SearchResults.tsx
@@ -31,9 +31,9 @@ import { useGeolocation } from 'libs/geolocation'
 import { useIsFalseWithDelay } from 'libs/hooks/useIsFalseWithDelay'
 import { plural } from 'libs/plural'
 import { SearchHit } from 'libs/search'
+import { useOpacityTransition } from 'ui/animations/helpers/useOpacityTransition'
 import { ButtonSecondary } from 'ui/components/buttons/ButtonSecondary'
 import { styledButton } from 'ui/components/buttons/styledButton'
-import { useHeaderTransition as useOpacityTransition } from 'ui/components/headers/animationHelpers'
 import { Li } from 'ui/components/Li'
 import { useModal } from 'ui/components/modals/useModal'
 import { GenericBanner } from 'ui/components/ModuleBanner/GenericBanner'
