+++ src/features/venue/components/VenueHeader/VenueHeader.tsx
@@ -7,7 +7,7 @@ import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
 import { useGoBack } from 'features/navigation/useGoBack'
 import { useShareVenue } from 'features/venue/helpers/useShareVenue'
 import { WebShareModal } from 'libs/share/WebShareModal'
-import { getAnimationState } from 'ui/components/headers/animationHelpers'
+import { getAnimationState } from 'ui/animations/helpers/getAnimationState'
 import { HeaderIcon } from 'ui/components/headers/HeaderIcon'
 import { useModal } from 'ui/components/modals/useModal'
 import { Spacer, Typo } from 'ui/theme'
