+++ src/ui/animations/helpers/getAnimationState.ts
@@ -1,15 +1,5 @@
-import { useRef } from 'react'
-import { Animated, Easing, NativeScrollEvent, NativeSyntheticEvent } from 'react-native'
+import { Animated, Easing } from 'react-native'
 import { DefaultTheme } from 'styled-components/native'
-
-import { getSpacing } from 'ui/theme'
-const HEIGHT_END_OF_TRANSITION = getSpacing(20)
-
-const interpolationConfig = {
-  inputRange: [0, HEIGHT_END_OF_TRANSITION],
-  outputRange: [0, 1],
-  extrapolate: 'clamp' as Animated.ExtrapolateType,
-}
 
 const iconBackgroundInterpolation = (theme: DefaultTheme) => ({
   inputRange: [0, 1],
@@ -39,20 +29,3 @@ export const getAnimationState = (
   },
   backgroundColor: headerTransition.interpolate(headerBackgroundInterpolation(theme)),
 })
-
-interface Props {
-  listener?: ({ nativeEvent }: NativeSyntheticEvent<NativeScrollEvent>) => void
-}
-
-export const useHeaderTransition = ({ listener }: Props = {}) => {
-  const headerScroll = useRef(new Animated.Value(0)).current
-
-  const onScroll = Animated.event([{ nativeEvent: { contentOffset: { y: headerScroll } } }], {
-    useNativeDriver: false,
-    listener,
-  })
-
-  const headerTransition = headerScroll.interpolate(interpolationConfig)
-
-  return { headerTransition, onScroll }
-}
